fixed ServerUser entity definition
