+++ database/entity/index.ts
@@ -4,6 +4,7 @@ import { LoginEmailOptIn } from './LoginEmailOptIn'
 import { LoginUser } from './LoginUser'
 import { LoginUserBackup } from './LoginUserBackup'
 import { Migration } from './Migration'
+import { ServerUser } from './ServerUser'
 import { Transaction } from './Transaction'
 import { TransactionCreation } from './TransactionCreation'
 import { TransactionSendCoin } from './TransactionSendCoin'
@@ -18,6 +19,7 @@ export const entities = [
   LoginUser,
   LoginUserBackup,
   Migration,
+  ServerUser,
   Transaction,
   TransactionCreation,
   TransactionSendCoin,
